requirements.txt
duckduckgo_search==5.3.0
llama_index==0.10.34
llama-index-embeddings-huggingface
llama-index-llms-anthropic
llama-index-llms-groq
llama_parse==0.4.2
phidata==2.3.84
python-dotenv==1.0.1
streamlit==1.33.0
tabulate
yfinance==0.2.38
